Add user projects page with shared projects component
