+++ packages/web/src/layouts/MainLayout.tsx
@@ -1,23 +1,78 @@
-import { Outlet } from 'react-router-dom'
+import { Outlet, Link, useLocation } from 'react-router-dom'
 import { UserMenu } from '@/components/UserMenu'
-import { Link } from 'react-router-dom'
+import { Home, FolderKanban } from 'lucide-react'
+import { cn } from '@/lib/utils'
 
 export function MainLayout() {
+  const location = useLocation()
+
+  const navigation = [
+    { name: 'Home', href: '/', icon: Home },
+    { name: 'Projects', href: '/projects', icon: FolderKanban },
+  ]
+
   return (
     <div className="min-h-screen bg-gray-50 dark:bg-gray-900">
       <header className="bg-white dark:bg-gray-800 shadow-sm">
         <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
           <div className="flex justify-between items-center py-4 lg:py-6">
-            <Link
-              to="/"
-              className="text-xl lg:text-2xl font-bold text-gray-900 dark:text-white"
-            >
-              SpecifAI
-            </Link>
+            <div className="flex items-center space-x-8">
+              <Link
+                to="/"
+                className="text-xl lg:text-2xl font-bold text-gray-900 dark:text-white"
+              >
+                SpecifAI
+              </Link>
+              
+              {/* Navigation Menu */}
+              <nav className="hidden md:flex space-x-1">
+                {navigation.map((item) => {
+                  const isActive = location.pathname === item.href
+                  return (
+                    <Link
+                      key={item.name}
+                      to={item.href}
+                      className={cn(
+                        'flex items-center space-x-2 px-3 py-2 rounded-md text-sm font-medium transition-colors',
+                        isActive
+                          ? 'bg-gray-900 text-white dark:bg-gray-100 dark:text-gray-900'
+                          : 'text-gray-900 hover:bg-gray-100 dark:text-gray-100 dark:hover:bg-gray-700'
+                      )}
+                    >
+                      <item.icon className="w-4 h-4" />
+                      <span>{item.name}</span>
+                    </Link>
+                  )
+                })}
+              </nav>
+            </div>
+            
             <div className="flex items-center space-x-2 lg:space-x-4">
               <UserMenu />
             </div>
           </div>
+          
+          {/* Mobile Navigation */}
+          <nav className="md:hidden flex space-x-1 pb-3">
+            {navigation.map((item) => {
+              const isActive = location.pathname === item.href
+              return (
+                <Link
+                  key={item.name}
+                  to={item.href}
+                  className={cn(
+                    'flex items-center space-x-2 px-3 py-2 rounded-md text-sm font-medium transition-colors flex-1 justify-center',
+                    isActive
+                      ? 'bg-gray-900 text-white dark:bg-gray-100 dark:text-gray-900'
+                      : 'text-gray-900 hover:bg-gray-100 dark:text-gray-100 dark:hover:bg-gray-700'
+                  )}
+                >
+                  <item.icon className="w-4 h-4" />
+                  <span>{item.name}</span>
+                </Link>
+              )
+            })}
+          </nav>
         </div>
       </header>
