+++ packages/web/src/pages/UserProjectsPage.tsx
@@ -1,0 +1,9 @@
+import { ProjectsPageContent } from '@/components/ProjectsPageContent'
+
+/**
+ * Projects page for regular users
+ * Shows only their assigned issues in a read-only view
+ */
+export function UserProjectsPage() {
+  return <ProjectsPageContent isReadOnly={true} userView={true} />
+}
